Squashed commit of the following:
    Endless scroll

    Ошибки

    Избранное, друзья, скрытые записи

    user feed

+++ app/src/main/java/ru/taaasty/service/Feeds.java
@@ -13,7 +13,7 @@ import rx.Observable;
  */
 public interface Feeds {
     @GET("/feeds/live.json")
-    Observable<Feed> getLiveFeed(@Query("since_entry_id") Integer sinceEntryId,
+    Observable<Feed> getLiveFeed(@Query("since_entry_id") Long sinceEntryId,
                      @Query("limit") Integer limit);
